chore(test): deduplicate property test helpers
Move sortPts/transformCircle/transformPoint into tests/fixtures/geom.ts and reuse; tests green.
