+++ tests/property/circle.properties.test.ts
@@ -1,6 +1,7 @@
 import { test, fc } from "@fast-check/vitest";
 import { circleCircleIntersection } from "../../src/geom/circle";
 import type { Circle, IntersectResult, Vec } from "../../src/geom/types";
+import { sortPts, transformCircle, transformPoint } from "../fixtures/geom";
 
 const range = 20;
 
@@ -18,27 +19,6 @@ function isTwoOrTangent(r: IntersectResult): r is IntersectResult & { points: Ve
     return (r.kind === "two" || r.kind === "tangent") && Array.isArray(r.points);
 }
 
-function sortPts(pts: Vec[]): Vec[] {
-    return [...pts].sort((p, q) => (p.x === q.x ? p.y - q.y : p.x - q.x));
-}
-
-function transformCircle(s: number, angle: number, tx: number, ty: number, c: Circle): Circle {
-    const ca = Math.cos(angle);
-    const sa = Math.sin(angle);
-    const x = c.c.x;
-    const y = c.c.y;
-    const xr = s * (ca * x - sa * y) + tx;
-    const yr = s * (sa * x + ca * y) + ty;
-    return { c: { x: xr, y: yr }, r: Math.abs(s) * c.r };
-}
-
-function transformPoint(s: number, angle: number, tx: number, ty: number, p: Vec): Vec {
-    const ca = Math.cos(angle);
-    const sa = Math.sin(angle);
-    const xr = s * (ca * p.x - sa * p.y) + tx;
-    const yr = s * (sa * p.x + ca * p.y) + ty;
-    return { x: xr, y: yr };
-}
 
 describe("circleCircleIntersection properties (fast-check)", () => {
     test.prop([circleArb, circleArb])("points satisfy both circle equations (two/tangent)", (A, B) => {
